fod_common/authentication.py
import os
from typing import Annotated

from fastapi import Depends, HTTPException, status
from fastapi.security import OAuth2PasswordBearer
from jose import JWTError, jwt

SECRET_KEY = "srgghwerhw4ehwrrtgrwtgrt23453t243gb5tg45" #TODO: config or env
ALGORITHM = "HS256"

oauth2_bearer = OAuth2PasswordBearer(tokenUrl="http://authentication-service:80/auth/requestToken")

async def get_current_user(token: str = Depends(oauth2_bearer)):
    try:
        payload = jwt.decode(token, SECRET_KEY, algorithms=[ALGORITHM])
        username = payload.get("sub")
        user_id = payload.get("id")
        if not username or not user_id:
            raise HTTPException(status_code=status.HTTP_401_UNAUTHORIZED, detail="Could not validate user.")
        
        return {"username": username, "user_id": user_id}
    except JWTError:
        raise HTTPException(status_code=status.HTTP_401_UNAUTHORIZED, detail="Could not validate user.")
    
user_dependency = Annotated[dict, Depends(get_current_user)]