fod_common/requirements.txt
tortoise-orm[asyncpg]
fastapi
python-jose[cryptography]